Move formatGeneratedAt to insight_constants.ts and add as const
- Move formatGeneratedAt function from summary.tsx to insight_constants.ts
  to colocate it with other insight-related constants
- Add `as const satisfies` to impactBadgeColors and impactLabels for
  better type narrowing while preserving type safety
- Use formatGeneratedAt in summary.tsx render column instead of
  moment().fromNow() to restore i18n-aware relative time formatting
- Remove unused moment import from summary.tsx
